# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WebGames API",
        "description": "Documentación de la API para la gestión de WebGames",
        "contact": {
            "email": "soporte@webgames.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Obtener todos los comentarios paginados",
                "operationId": "2e2597016871fa5dd2bdfccf912edc41",
                "responses": {
                    "200": {
                        "description": "Lista de comentarios paginados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Comment"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Crear un nuevo comentario",
                "operationId": "c6ce2c222c8502aadd6a287105e49307",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "post_id",
                                    "body"
                                ],
                                "properties": {
                                    "post_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "Este es un comentario de prueba"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comentario creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/comments/{comment}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Obtener un comentario específico con sus respuestas",
                "operationId": "884e61ac3f4a0c4eb6d52f68929b4fc9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del comentario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comentario obtenido con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comentario no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Actualizar un comentario",
                "operationId": "2b73d856972689aa0e0d9223d46011ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del comentario a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "body"
                                ],
                                "properties": {
                                    "body": {
                                        "type": "string",
                                        "example": "Comentario actualizado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comentario actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado para actualizar este comentario"
                    },
                    "404": {
                        "description": "Comentario no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Eliminar un comentario",
                "operationId": "3d590dba6524ead27a250d196bc2729b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del comentario a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Comentario eliminado correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado para eliminar este comentario"
                    },
                    "404": {
                        "description": "Comentario no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Comment": {
                "title": "Comment",
                "description": "Estructura del comentario en la API",
                "required": [
                    "id",
                    "body",
                    "published_at",
                    "post_id",
                    "user_id",
                    "category_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID único del comentario",
                        "type": "integer",
                        "example": 1
                    },
                    "body": {
                        "description": "Contenido del comentario",
                        "type": "string",
                        "example": "Este es un comentario de prueba"
                    },
                    "published_at": {
                        "description": "Fecha y hora en que se publicó el comentario",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-21T13:00:20.000000Z"
                    },
                    "parent": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Comment"
                            }
                        ],
                        "nullable": true,
                        "description": "Comentario padre si es una respuesta"
                    },
                    "replies": {
                        "description": "Lista de respuestas a este comentario",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Comment"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}